+++ src/auditwheel.rs
@@ -95,7 +95,7 @@ pub fn auditwheel_rs(path: &Path) -> Result<(), AuditWheelError> {
     for dep in deps {
         // I'm not 100% what exactely this line does, but auditwheel also seems to skip
         // everything with ld-linux in its name
-        if dep == "ld-linux-x86-64.so.2" {
+        if dep == "ld-linux-x86-64.so.2" || dep == "ld-linux.so.2" {
             continue;
         }
         if !MANYLINUX1.contains(&dep.as_str()) {