+++ Changelog.md
@@ -15,7 +15,7 @@ get_fourtytwo-2.0.1-cp36-cp36m-manylinux1_x86_64.whl | 2,8M | 771K | 102K
 hello_world-0.1.0-py2.py3-none-manylinux1_x86_64.whl | 3,9M | 1,1M | 180K
 points-0.1.0-py2.py3-none-manylinux1_x86_64.whl | 2,8M | 752K | 85K
 
-### Added 
+### Added
 
  * `--strip` by ijl [#7](https://github.com/PyO3/pyo3-pack/pull/7)
 
@@ -23,6 +23,10 @@ points-0.1.0-py2.py3-none-manylinux1_x86_64.whl | 2,8M | 752K | 85K
 
  * Renamed `--bindings-crate` to `--bindings`
  * Use deflate compression for zips by ijl [#6](https://github.com/PyO3/pyo3-pack/pull/6)
+
+### Fixed
+
+ * `--target` is now actually used for the wheel compatibility tag
 
 ## [0.3.5] - 2018-09-20